% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FixelArray.R
\name{FixelArray}
\alias{FixelArray}
\title{Create a fixel array data type for fixel-based analysis in R}
\usage{
FixelArray(filepath, scalar_types = c("FD"))
}
\arguments{
\item{filepath}{The HDF5 file created with Docker}

\item{scalar_types}{A list of expected scalars, can be extracted from your cohort file}
}
\value{
A FixelArray object
}
\description{
Create a fixel array data type for fixel-based analysis in R
}
